216 update teal.code version in dependencies (#224)
Closes #216 

qenv replaced chunks in this PR
https://github.com/insightsengineering/teal.goshawk/pull/203/files -
this is a backend change and this was not exposed in vignettes and not
exposed in examples. I reviewed examples and vignettes in the last 2
weeks while working on #223 and #221 and I run those examples multiple
times recently. Most of the things marked in #216 were not needed to be
checked, besides the teal.code version bump.