284 add callback for `DT::datable` to hide AJAX warnings (#310)
Solution for #284 that do not pop-up AJAX warnings from JavaScript
DataTable when computations happens too often and rendering is behind
the state of the data.

This adds a callback behavior for `DT::datatable`. We also tested the
global setup of this behavior in teal but we were unable to include
javascript in a way that it is executed after the object is created.
Other modules in other packages don't really need that as they have
better reactivity handling. The modules in teal.goshawk have very much
reactivitiy and could be refactored one day but that's a bigger project.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>